+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## [Unreleased]
+## [0.0.2] - 2026-02-26
 
 ### Added
 
 - `User`, `EmailUser`, `EmailUserManager`
+- `TimezoneMixin` and `TimezoneMiddleware`
 
 ## [0.0.1] - 2026-02-26